Prevent Grant Duration delete if it is associated to one or more Grants

+++ app/src/main/java/io/unityfoundation/dds/permissions/manager/model/grantduration/GrantDurationService.java
@@ -20,6 +20,7 @@ import io.micronaut.http.HttpStatus;
 import io.micronaut.http.MutableHttpResponse;
 import io.unityfoundation.dds.permissions.manager.ResponseStatusCodes;
 import io.unityfoundation.dds.permissions.manager.exception.DPMException;
+import io.unityfoundation.dds.permissions.manager.model.applicationgrant.ApplicationGrantRepository;
 import io.unityfoundation.dds.permissions.manager.model.grantduration.dto.GrantDurationDTO;
 import io.unityfoundation.dds.permissions.manager.model.grantduration.dto.CreateGrantDurationDTO;
 import io.unityfoundation.dds.permissions.manager.model.group.Group;
@@ -39,12 +40,14 @@ public class GrantDurationService {
 
     private final GrantDurationRepository grantDurationRepository;
     private final GroupRepository groupRepository;
+    private final ApplicationGrantRepository grantRepository;
     private final SecurityUtil securityUtil;
     private final GroupUserService groupUserService;
 
-    public GrantDurationService(GrantDurationRepository grantDurationRepository, GroupRepository groupRepository, SecurityUtil securityUtil, GroupUserService groupUserService) {
+    public GrantDurationService(GrantDurationRepository grantDurationRepository, GroupRepository groupRepository, ApplicationGrantRepository grantRepository, SecurityUtil securityUtil, GroupUserService groupUserService) {
         this.grantDurationRepository = grantDurationRepository;
         this.groupRepository = groupRepository;
+        this.grantRepository = grantRepository;
         this.securityUtil = securityUtil;
         this.groupUserService = groupUserService;
     }
@@ -181,8 +184,16 @@ public class GrantDurationService {
 
         checkExistenceAndAdminAuthorization(grantDurationOptional);
 
+        if (hasGrantAssociations(grantDurationOptional.get())) {
+            throw new DPMException(ResponseStatusCodes.GRANT_DURATION_HAS_ONE_OR_MORE_GRANT_ASSOCIATIONS, HttpStatus.BAD_REQUEST);
+        }
+
         grantDurationRepository.delete(grantDurationOptional.get());
         return HttpResponse.noContent();
+    }
+
+    private boolean hasGrantAssociations(GrantDuration grantDuration) {
+        return grantRepository.existsByGrantDuration(grantDuration);
     }
 
     public GrantDurationDTO createDTO(GrantDuration grantDuration) {
